% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_spread.R
\name{catch_spread}
\alias{catch_spread}
\title{Spread raw catch data to appropriate knots}
\usage{
catch_spread(catch = NULL, knots = NULL, area = NULL, crs_input = NULL)
}
\arguments{
\item{catch}{sf object with columns Catch, Year, geometry}

\item{knots}{kmeans object obtained from setup_mesh function}

\item{area}{area covered by each knot as obtained from setup_pred_grid function}

\item{crs_input}{Desired crs, currently only UTM is accepted}
}
\value{
List containing properly formatted catch, either sums per at each knot or density (catch/km^2) at each knot
}
\description{
Spread raw catch data to appropriate knots
}
