% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{Organize a dataframe containing the individual observations to work with TMB model}
\usage{
data_setup(
  data = NULL,
  growths = NULL,
  catch = NULL,
  model = NULL,
  mesh = NULL,
  bound = NULL,
  obs_mort = FALSE,
  prior = FALSE,
  fix_m = 0.1,
  mult_qI = FALSE,
  spat_approach = NULL,
  knot_obj = NULL,
  knot_area = NULL,
  separate_R_aniso = T
)
}
\arguments{
\item{data}{list: containing individual observations: I for commercial size, IR for recruit size, Year for year, if spatial: geometry to show object is an sf object, optional: L for clappers, N for total numbers.}

\item{growths}{dataframe: one column for commercial size growth rate g, one for recruit growth rate gR}

\item{catch}{if non-spatial: vector of yearly catches, if spatial: data frame of catches with knots as rows and years as columns}

\item{model}{character: choice of model, options: SEBDAM or TLM}

\item{mesh}{INLA mesh: used for spatial approach, if non-spatial leave null, obtained from setup_mesh function}

\item{bound}{sf object: used to get triangles for barrier model, if not barrier model leave null}

\item{obs_mort}{boolean: true if using observations of mortality (e.g. clappers), false if fixing mortality}

\item{prior}{boolean: true if using a prior distribution to inform q_I, false to freely estimate}

\item{fix_m}{double: value used for fixing mortality when not using any observations, ignore if using observations}

\item{mult_qI}{boolean: if not using spatial approach, ignore, if using spatial: true when using a separate catchability at each knot, false for using a single overall catchability}

\item{spat_approach}{character: if using TLM, ignore, if using SEBDAM, choice of spatial model: "spde" for normal SPDE, "spde_aniso" for SPDE approach with geometric anisotropy, and "barrier" for barrier model}

\item{knot_obj}{kmeans: objects that set knot for spatial approach, can be ignored if using TLM, obtained from setup_mesh function}

\item{knot_area}{dataframe: obtained from setup_pred_grid function, used to know area used for each knot in spatial approach, ignore if using TLM}

\item{separate_R_aniso}{boolean: if using SPDE with geometric anisotropy, set to F is want biomass and recruitment to have same anisotropy}
}
\value{
Properly formatted lists for model fitting
}
\description{
Organize a dataframe containing the individual observations to work with TMB model
}
