% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_utm_code}
\alias{find_utm_code}
\title{Find the EPGS code for UTM zone code of given WGS 84 longitude and latitude}
\usage{
find_utm_code(lonlat)
}
\arguments{
\item{lonlat}{Vector of WGS 84 with entries of longitude and latitude}
}
\value{
EPSG code UTM zone number
}
\description{
Find the EPGS code for UTM zone code of given WGS 84 longitude and latitude
}
