% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit TLM or SEBDAM to fisheries data}
\usage{
fit_model(
  tmb_obj,
  optim = "nlminb",
  control = NULL,
  optim_method = NULL,
  cores = parallel::detectCores(),
  bias.correct = F
)
}
\arguments{
\item{tmb_obj}{List: properly formatted data and parameter starting values from data_setup function}

\item{optim}{Character: choice of optimizers to use, choices are nlminb, optimr, or parallel (use multiple cores)}

\item{control}{list: control optimizer parameters}

\item{optim_method}{if using optimr, choice of optimization method}

\item{cores}{detects cores for parallel optimization unless otherwise specified}

\item{bias.correct}{for TMB::sdreport, if want to apply bias correction}
}
\value{
Object containing modelling output
}
\description{
Fit TLM or SEBDAM to fisheries data
}
