% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_params.R
\name{fix_param}
\alias{fix_param}
\title{Fix specific parameters to user-specified values}
\usage{
fix_param(obj = NULL, pars = NULL)
}
\arguments{
\item{obj}{TMB object ready to be fit, obtained from data_setup; same object that would be given to fit_model}

\item{pars}{list:}
}
\value{

}
\description{
Fix specific parameters to user-specified values
}
