% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Obtain data frames of parameter and their standard errors}
\usage{
get_parameters(return_obj)
}
\arguments{
\item{return_obj}{Fitted modelling output for fit_model function}
}
\value{
dataframe: parameter estimates and standard errors
}
\description{
Obtain data frames of parameter and their standard errors
}
