% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_processes.R
\name{get_processes}
\alias{get_processes}
\title{Obtain data frame of underlying processes and standard errors}
\usage{
get_processes(return_obj)
}
\arguments{
\item{return_obj}{Fitted modelling output from fit_model function}
}
\value{
dataframe: underlying processes and standard errors
}
\description{
Obtain data frame of underlying processes and standard errors
}
