% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_triangles}
\alias{get_triangles}
\title{Finding INLA mesh vertices located on land for barrier model}
\usage{
get_triangles(mesh, bound)
}
\arguments{
\item{mesh}{INLA mesh, preferably obtained from setup_mesh function}

\item{bound}{sf objects: modelling bounds including islands/land for barrier model}
}
\value{
object for barrier model indicating triangles/vertices on land
}
\description{
Finding INLA mesh vertices located on land for barrier model
}
