% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rand_int_vect}
\alias{rand_int_vect}
\title{Randomly create vector of integers that sum up to specified total}
\usage{
rand_int_vect(min, max, n_grp, total)
}
\arguments{
\item{min}{minimum number desired}

\item{max}{max number desired}

\item{n_grp}{number of integers to be simulated}

\item{total}{sum of integers desired}
}
\value{
vector of integers
}
\description{
Randomly create vector of integers that sum up to specified total
}
