% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mesh.R
\name{setup_mesh}
\alias{setup_mesh}
\title{Setting up the predictive knots and mesh based on the knots}
\usage{
setup_mesh(
  data = NULL,
  seed = NULL,
  nknot = 25,
  model_bound = NULL,
  crs_input = NULL,
  from_sim = FALSE,
  max.edge = c(8, 30),
  cutoff = 2,
  ...
)
}
\arguments{
\item{data}{sf object containing data and locations of all tows}

\item{seed}{specify seed number for reproducibility}

\item{nknot}{integer: number of knots wanted for number of knots, default set at 25}

\item{model_bound}{sf object for boundaries of modelling area}

\item{crs_input}{Specify crs that want to model in, but not tested for anything that is not UTM at the moment}

\item{from_sim}{boolean: true if using simulated data, false if using real data}

\item{max.edge}{inla.mesh.2d control}

\item{cutoff}{inla.mesh.2d control}

\item{...}{other parameters fed to inla.mesh.2d}
}
\value{

}
\description{
Setting up the predictive knots and mesh based on the knots
}
