% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_TLM_params.R
\name{simul_TLM_params}
\alias{simul_TLM_params}
\title{Setup parameters for simulations using TLM}
\usage{
simul_TLM_params(
  simul_data_obj = NULL,
  sigma_tau = 0.1,
  sigma_phi = 0.1,
  sigma_epsilon = 0.1,
  sigma_upsilon = 0.1,
  q_R = 0.2,
  q_I = 0.45,
  p_I = 0.9,
  p_IR = 0.4,
  sigma_m = 0.1,
  S = 0.4
)
}
\arguments{
\item{simul_data_obj}{List: object obtained from simulate_obj function}

\item{sigma_tau}{double: biomass process variance parameter}

\item{sigma_phi}{double: recruitment process variance parameter}

\item{sigma_epsilon}{double: biomass observation variance parameter}

\item{sigma_upsilon}{double: recruitment observation variance parameter}

\item{q_R}{double: recruitment catchability}

\item{q_I}{double: biomass catchability}

\item{p_I}{double: probability of biomass capture}

\item{p_IR}{double: probability of recruit capture}

\item{sigma_m}{double: natural mortality process variance parameter}

\item{S}{double: clapper catchability}
}
\value{
Formatted list for simulation purposes using TLM
}
\description{
Setup parameters for simulations using TLM
}
