% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_area.R
\name{simulate_area}
\alias{simulate_area}
\title{Create an area, random locations, knots and corresponding mesh for the purpose of simulations}
\usage{
simulate_area(
  n_obs = NULL,
  n_knots = NULL,
  seed = NULL,
  ...,
  x_coord = c(175, 225, 225, 175),
  y_coord = c(175, 175, 225, 225)
)
}
\arguments{
\item{n_obs}{Number of simulated observations}

\item{n_knots}{Number of desired knots}

\item{seed}{Seed for reproducibility of knots}

\item{...}{Given to sf::st_sample for different types of random samplings, and to INLA::inla.mesh.2d for additional parameters, especially if not using defaults}

\item{x_coord}{x coordinates for creation of simulation area}

\item{y_coord}{y coordinates for creation of simulation area}
}
\value{
list containing sf object for simulation area, kmeans object for knots, inla mesh, and sfc object with location of every simulated points
}
\description{
Create an area, random locations, knots and corresponding mesh for the purpose of simulations
}
