% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data and underlying processes using either TLM or SEBDAM}
\usage{
simulate_data(
  simul_obj = NULL,
  seed = NULL,
  format = "formatted",
  sim_obs_loc = NULL
)
}
\arguments{
\item{simul_obj}{List: object for simulations obtained from simul_TLM_params or simul_SEBDAM_params function}

\item{seed}{integer: to set the seed for data simulation}

\item{format}{character: choose format of output, "raw" simply returns the list directly created by obj$simulate, "formatted" returns an organized list separating data, growths, catches, and simulated processes}

\item{sim_obs_loc}{sfc objecT: tow locations obtained from simul_area function, only needed if using SEBDAM}
}
\value{
List containing simulated data and processes
}
\description{
Simulate data and underlying processes using either TLM or SEBDAM
}
