% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_obj.R
\name{simulate_obj}
\alias{simulate_obj}
\title{Create object used for simulations using SEBDAM or TLM}
\usage{
simulate_obj(
  model = NULL,
  n_years = 1L,
  obs_mort = FALSE,
  n_obs = NULL,
  even_spread = TRUE,
  fix_m = 0.1,
  prior_q = FALSE,
  catch_spread = NULL,
  simul_area_obj = NULL,
  mult_qI = FALSE,
  spat_approach = "spde",
  separate_R_aniso = TRUE,
  bound = NULL
)
}
\arguments{
\item{model}{Character: Choose which model to use, options are TLM or SEBDAM}

\item{n_years}{Integer: Number of years desired}

\item{obs_mort}{Boolean: True to simulate clapper observations and processes, false to fix natural mortality}

\item{n_obs}{Integer: required for TLM and ignored for SEBDAM (as number is chosen in simulate_area function), choose total number of observations}

\item{even_spread}{Boolean: choice to spread observations evenly across years, or randomly split them up}

\item{fix_m}{Double: only matters for TLM, if obs_mort is FALSE, choice of value to fix the natural mortality (default of 0.1)}

\item{prior_q}{boolean: TRUE to use a prior distribution on q_I in later fitting processes}

\item{catch_spread}{Character: ignored if using TLM, choice of how to simulate catches, choices are prop, aggregated or aggregated_extra}

\item{simul_area_obj}{List: ignored if using TLM, object obtained from running simulate_area function for SEBDAM}

\item{mult_qI}{boolean: ignored if using TLM, choice of using a single q_I (FALSE) or one per knot (TRUE)}

\item{spat_approach}{character: ignored if using TLM, choice of spatial approach, options are spde, spde_aniso and barrier}

\item{separate_R_aniso}{boolean: ignored if using TLM or using spde or barrier spatial approaches, choice of having separate anisotropy parameters for R or the same as for B}

\item{bound}{sf object: boundaries of modelling only when using barrier model, must includes islands}
}
\value{
list: object used for simulations in simulate_data function
}
\description{
Create object used for simulations using SEBDAM or TLM
}
