% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SEBDAM_params.R
\name{simulate_SEBDAM_params}
\alias{simulate_SEBDAM_params}
\title{Setup parameters for simulations using SEBDAM}
\usage{
simulate_SEBDAM_params(
  simul_data_obj = NULL,
  sigma_epsilon = 0.1,
  sigma_upsilon = 0.1,
  R0 = 100,
  B0 = 500,
  m0 = 0.1,
  p_I = 0.9,
  p_IR = 0.4,
  qR = 0.2,
  range_B = 40,
  range_R = 30,
  sigma_B = 0.1,
  sigma_R = 0.1,
  H_input_B = c(-1.2, -0.5),
  H_input_R = c(-1, -0.8),
  qI = 0.45,
  range_m = 20,
  sigma_m = 0.1,
  H_input_m = c(-0.5, 0.4),
  S = 0.4
)
}
\arguments{
\item{simul_data_obj}{List: object obtained from simulate_obj function}

\item{sigma_epsilon}{double: biomass observation variance parameter}

\item{sigma_upsilon}{double: recruitment observation variance parameter}

\item{R0}{double: recruitment density mean value in initial year}

\item{B0}{double: biomass density mean value in initial year}

\item{m0}{double: if using clapper approach, natural mortality mean value in initial year, if not then value at which natural mortality will be fixed}

\item{p_I}{double: biomass probability of capture}

\item{p_IR}{double: recruit probability of capture}

\item{qR}{double: recruitment catchability}

\item{range_B}{double: range at which two points become decorrelated for biomass}

\item{range_R}{double: range at which two points become decorrelated for recruitment}

\item{sigma_B}{double: marginal variance of biomass random fields}

\item{sigma_R}{double: marginal variance of recruit random fields}

\item{H_input_B}{double: anisotropy parameters for biomass random fields}

\item{H_input_R}{double: anisotropy parameters for recruitment random fields}

\item{qI}{double: biomass catchability}

\item{range_m}{double: range at which two points become decorrelated for natural mortality}

\item{sigma_m}{double: marginal variance of natural mortality random fields}

\item{H_input_m}{double: anisotropy parameters for natural mortality random fields}

\item{S}{double: clapper catchability}
}
\value{
Formatted list for simulation purposes using SEBDAM
}
\description{
Setup parameters for simulations using SEBDAM
}
